requirements.txt
python-Levenshtein==0.11.2
beautifulsoup4==4.4.1
requests==2.4.0
paramiko==1.16.0
jsonpath-rw==1.4.0
selenium==2.53.2
responses==0.5.1
